Simplifying the LimitRate operator
