+++ reactor-core/src/test/java/reactor/core/publisher/FluxPrefetchTest.java
@@ -16,7 +16,13 @@
 
 package reactor.core.publisher;
 
+import java.time.Duration;
 import java.util.ArrayList;
+import java.util.Collections;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.LongAdder;
 import java.util.stream.Stream;
 
@@ -24,6 +30,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
+import org.reactivestreams.Subscription;
 import reactor.core.Fuseable;
 import reactor.test.StepVerifier;
 import reactor.test.subscriber.AssertSubscriber;
@@ -228,5 +235,98 @@ public class FluxPrefetchTest {
 		    .thenRequest(Long.MAX_VALUE)
 		    .expectNextCount(sourceSize)
 		    .verifyComplete();
+	}
+
+	// LimitRate Tests
+	@Test
+	public void limitRate() {
+		List<Long> upstreamRequests = new LinkedList<>();
+		List<Long> downstreamRequests = new LinkedList<>();
+		Flux<Integer> source = Flux.range(1, 400)
+		                           .doOnRequest(upstreamRequests::add)
+		                           .doOnRequest(r -> System.out.println(
+				                           "upstream request of " + r))
+		                           .hide()
+		                           .limitRate(40)
+		                           .doOnRequest(downstreamRequests::add)
+		                           .doOnRequest(r -> System.out.println(
+				                           "downstream request of " + r));
+
+		AssertSubscriber<Integer> ts = AssertSubscriber.create(400);
+		source.subscribe(ts);
+		ts.await(Duration.ofMillis(100))
+		  .assertComplete();
+
+		assertThat(downstreamRequests.size()).as("downstream number of requests").isOne();
+		assertThat(downstreamRequests.get(0)).as("unique request").isEqualTo(400L);
+		long total = 0L;
+		for (Long requested : upstreamRequests) {
+			total += requested;
+			//30 is the optimization that eagerly prefetches when 3/4 of the request has been served
+			assertThat(requested).as("rate limit check").isIn(40L, 30L);
+		}
+		assertThat(total).as("upstream total request")
+		                 .isGreaterThanOrEqualTo(400L)
+		                 .isLessThan(440L);
+	}
+
+	@Test
+	public void limitRateWithCloseLowTide() {
+		List<Long> rebatchedRequest = Collections.synchronizedList(new ArrayList<>());
+
+		final Flux<Integer> test = Flux
+				.range(1, 14)
+				.hide()
+				.doOnRequest(rebatchedRequest::add)
+				.limitRate(10,8);
+
+		StepVerifier.create(test, 14)
+		            .expectNextCount(14)
+		            .verifyComplete();
+
+		assertThat(rebatchedRequest)
+				.containsExactly(10L, 8L);
+	}
+
+	@Test
+	public void limitRateWithVeryLowTide() {
+		List<Long> rebatchedRequest = Collections.synchronizedList(new ArrayList<>());
+
+		final Flux<Integer> test = Flux
+				.range(1, 14)
+				.hide()
+				.doOnRequest(rebatchedRequest::add)
+				.limitRate(10,2);
+
+		StepVerifier.create(test, 14)
+		            .expectNextCount(14)
+		            .verifyComplete();
+
+		assertThat(rebatchedRequest)
+				.containsExactly(10L, 2L, 2L, 2L, 2L, 2L, 2L, 2L);
+	}
+
+	@Test
+	public void limitRateDisabledLowTide() throws InterruptedException {
+		LongAdder request = new LongAdder();
+		CountDownLatch latch = new CountDownLatch(1);
+
+		Flux.range(1, 99).hide()
+		    .doOnRequest(request::add)
+		    .limitRate(10, 0)
+		    .subscribe(new BaseSubscriber<Integer>() {
+			    @Override
+			    protected void hookOnSubscribe(Subscription subscription) {
+				    request(100);
+			    }
+
+			    @Override
+			    protected void hookFinally(SignalType type) {
+				    latch.countDown();
+			    }
+		    });
+
+		assertThat(latch.await(1, TimeUnit.SECONDS)).as("took less than 1s").isTrue();
+		assertThat(request.sum()).as("request not compounded").isEqualTo(100L);
 	}
 }
